refactor(control-plane): split grpc_handler.go by domain
Split grpc_handler.go (1002 lines) into three focused files:
- grpc_handler.go (573 lines): struct, constructor, handler, step helpers
- grpc_handler_mappers.go (291 lines): proto-to-domain mappers
- grpc_handler_phase.go (156 lines): phase status tracking

Remove from knownOversizedFiles in architecture tests.
